Refactor ETL job orchestration
