+++ backend/app/services.py
@@ -4,18 +4,18 @@ import sqlite3
 
 from fastapi import BackgroundTasks
 
-from . import etl, schemas
+from . import etl_runner, schemas
 
 
 def start_refresh(
     background_tasks: BackgroundTasks,
     _payload: schemas.RefreshTriggerPayload | None = None,
 ) -> schemas.RefreshResponse:
-    background_tasks.add_task(etl.start_etl_job)
-    return schemas.RefreshResponse(state=etl.STATE_RUNNING)
+    background_tasks.add_task(etl_runner.start_etl_job)
+    return schemas.RefreshResponse(state=etl_runner.STATE_RUNNING)
 
 
 def refresh_status(conn: sqlite3.Connection) -> schemas.RefreshStatusResponse:
-    status = etl.get_last_status(conn)
+    status = etl_runner.get_last_status(conn)
     payload = status.model_dump() if hasattr(status, "model_dump") else status.dict()
     return schemas.RefreshStatusResponse(**payload)
